refactor: rename services/ingestion_client.py → services/ingestion.py
